turn on 64bit ObjectSize
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/libgphoto2@14028 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ camlibs/ptp2/ptp.h
@@ -790,7 +790,11 @@ struct _PTPObjectInfo {
 	uint32_t StorageID;
 	uint16_t ObjectFormat;
 	uint16_t ProtectionStatus;
-	uint32_t ObjectCompressedSize;
+	/* In the regular objectinfo this is 32bit,
+	 * but we keep the general object size here
+	 * that also arrives via other methods and so
+	 * use 64bit */
+	uint64_t ObjectCompressedSize;
 	uint16_t ThumbFormat;
 	uint32_t ThumbCompressedSize;
 	uint32_t ThumbPixWidth;
